Organize architecture
