+++ __tests__/unit/api.test.js
@@ -1,4 +1,4 @@
-import api from '../../services/api';
+import api from '../../src/app/services/api';
 
 describe('api with axios', () => {
   it('should return an instance of axios', () => {
